ADD: Sparse matrix mode

+++ src/global/misc/inc/MResponseMatrixAxis.h
@@ -58,9 +58,9 @@ class MResponseMatrixAxis
   //! Set the bin edges
   void SetBinEdges(vector<double> BinEdges) { m_BinEdges = BinEdges; }
   //! Set the axis in linear mode
-  void SetLinear(unsigned int NBins, double Min, double Max, double UnderFlowMin = g_DoubleNotDefined, double OverFlowMax = g_DoubleNotDefined);
+  void SetLinear(unsigned long NBins, double Min, double Max, double UnderFlowMin = g_DoubleNotDefined, double OverFlowMax = g_DoubleNotDefined);
   //! Set the axis in logarithmic mode
-  void SetLogarithmic(unsigned int NBins, double Min, double Max, double UnderFlowMin = g_DoubleNotDefined, double OverFlowMax = g_DoubleNotDefined);
+  void SetLogarithmic(unsigned long NBins, double Min, double Max, double UnderFlowMin = g_DoubleNotDefined, double OverFlowMax = g_DoubleNotDefined);
   
   //! The dimensions of the axis
   unsigned int GetDimension() const { return m_Dimension; }
@@ -69,15 +69,15 @@ class MResponseMatrixAxis
   bool IsLogarithmic() const { return m_IsLogarithmic; }
   
   //! Return the number of axis bins
-  virtual unsigned int GetNumberOfBins() const { if (m_BinEdges.size() <= 1) { return 0; } else { return m_BinEdges.size() - 1; } }
+  virtual unsigned long GetNumberOfBins() const { if (m_BinEdges.size() <= 1) { return 0; } else { return m_BinEdges.size() - 1; } }
   //! Return the axis bin given a axis value
   //! Can throw: MExceptionArbitrary, MExceptionValueOutOfBounds, MExceptionIndexOutOfBounds
   //! To avoid any exceptions, call InRange() first!
-  virtual unsigned int GetAxisBin(double X1, double X2 = g_DoubleNotDefined) const;
+  virtual unsigned long GetAxisBin(double X1, double X2 = g_DoubleNotDefined) const;
   
   //! Return the area of the given axis bin
   //! Can throw: MExceptionIndexOutOfBounds
-  virtual double GetArea(unsigned int Bin) const;
+  virtual double GetArea(unsigned long Bin) const;
   
   //! Test if the x-value is within the range of the axis - the second value in always ignored
   virtual bool InRange(double X1, double X2 = g_DoubleNotDefined) const;
@@ -99,7 +99,7 @@ class MResponseMatrixAxis
   
   //! Return the bin center(s) of the given axis bin
   //! Can throw: MExceptionIndexOutOfBounds
-  virtual vector<double> GetBinCenters(unsigned int Bin) const;
+  virtual vector<double> GetBinCenters(unsigned long Bin) const;
   
   
   //! Write the content to a stream
