+++ src/global/misc/inc/MResponseMatrixAxisSpheric.h
@@ -50,13 +50,13 @@ class MResponseMatrixAxisSpheric : public MResponseMatrixAxis
   virtual MResponseMatrixAxisSpheric* Clone() const;
   
   //! Set the axis in FISBEL mode with a longitude shift in degrees
-  void SetFISBEL(unsigned int NBins, double LongitudeShift = 0);
+  void SetFISBEL(unsigned long NBins, double LongitudeShift = 0);
   
 
   //! Return the number of axis bins
-  virtual unsigned int GetNumberOfBins() const { return m_Binner.GetNBins(); }
+  virtual unsigned long GetNumberOfBins() const { return m_Binner.GetNBins(); }
   //! Return the axis bin, given theta=latitude and phi=longitude in degrees
-  virtual unsigned int GetAxisBin(double Theta, double Phi) const;
+  virtual unsigned long GetAxisBin(double Theta, double Phi) const;
   
   //! Test if theta and phi (in degree) are within the range of the axis - the second value is required
   virtual bool InRange(double Theta, double Phi) const;
@@ -68,7 +68,7 @@ class MResponseMatrixAxisSpheric : public MResponseMatrixAxis
   virtual vector<double> Get1DBinEdges() { return vector<double>(); }
   
   //! Return the area of the given axis bin
-  virtual double GetArea(unsigned int Bin) const;
+  virtual double GetArea(unsigned long Bin) const;
   
   //! Return the axis bin edges for drawing --- those might be narrower than the real bins
   virtual vector<vector<double>> GetDrawingAxisBinEdges() const;
@@ -79,7 +79,7 @@ class MResponseMatrixAxisSpheric : public MResponseMatrixAxis
   
   //! Return the bin center(s) of the given axis bin, theta & phi 
   //! Can throw: MExceptionIndexOutOfBounds
-  virtual vector<double> GetBinCenters(unsigned int Bin) const;
+  virtual vector<double> GetBinCenters(unsigned long Bin) const;
   
   //! Return the bin center of all axis bins as vectors 
   //! Can throw: MExceptionIndexOutOfBounds
